move script files to /firestoreScript/
