+++ firestoreScript/databaseConstants.cjs
@@ -1,0 +1,9 @@
+const allCollections = [
+  "cards",
+  "illustrators",
+  "rarity",
+  "series",
+  "sets",
+  "type",
+];
+module.exports = { allCollections };
