+++ package.json
@@ -13,8 +13,8 @@
     "lint": "eslint .",
     "preview": "vite preview",
     "tailwind": "tailwindcss",
-    "export-data": "node scripts/exportCollections.cjs",
-    "import-data": "node scripts/importCollections.cjs"
+    "export-data": "node firestoreScript/exportCollections.cjs",
+    "import-data": "node firestoreScript/importCollections.cjs"
   },
   "dependencies": {
     "@floating-ui/react-dom": "^2.1.2",
